feat: fix loader util problem
