+++ src/loader/LoaderUtils.ts
@@ -3,12 +3,10 @@ import {
   SkeletonBinary,
   SkeletonData,
   SkeletonJson,
-  Texture,
-  TextureAtlas,
-  TextureFilter,
-  TextureWrap
+  TextureAtlas
 } from "@esotericsoftware/spine-core";
-import { AssetPromise, AssetType, Engine, Texture2D, TextureFilterMode, TextureWrapMode } from "@galacean/engine";
+import { AssetPromise, AssetType, Engine, Texture2D } from "@galacean/engine";
+import { AdaptiveTexture } from "./AdaptiveTexture";
 import { SpineResource } from "./SpineResource";
 
 /**
@@ -119,48 +117,4 @@ export function getBaseUrl(url: string): string {
   const parsedUrl = new URL(url);
   const basePath = parsedUrl.origin + parsedUrl.pathname;
   return basePath.endsWith("/") ? basePath : basePath.substring(0, basePath.lastIndexOf("/") + 1);
-}
-
-export class AdaptiveTexture extends Texture {
-  texture: Texture2D;
-  constructor(image: HTMLImageElement | ImageBitmap, texture: Texture2D) {
-    super(image);
-    this.texture = texture;
-    this.texture.generateMipmaps();
-  }
-
-  // rewrite getImage function, return galacean texture2D, then attachment can get size of texture
-  getImage(): any {
-    return this.texture;
-  }
-
-  setFilters(minFilter: TextureFilter, magFilter: TextureFilter) {
-    if (minFilter === TextureFilter.Nearest) {
-      this.texture.filterMode = TextureFilterMode.Point;
-    } else if (magFilter === TextureFilter.MipMapLinearLinear) {
-      this.texture.filterMode = TextureFilterMode.Trilinear;
-    } else {
-      this.texture.filterMode = TextureFilterMode.Bilinear;
-    }
-  }
-
-  setWraps(uWrap: TextureWrap, vWrap: TextureWrap) {
-    this.texture.wrapModeU = this._convertWrapMode(uWrap);
-    this.texture.wrapModeV = this._convertWrapMode(vWrap);
-  }
-
-  dispose() {}
-
-  private _convertWrapMode(wrap: TextureWrap): TextureWrapMode {
-    switch (wrap) {
-      case TextureWrap.ClampToEdge:
-        return TextureWrapMode.Clamp;
-      case TextureWrap.Repeat:
-        return TextureWrapMode.Repeat;
-      case TextureWrap.MirroredRepeat:
-        return TextureWrapMode.Mirror;
-      default:
-        throw new Error("Unsupported texture wrap mode.");
-    }
-  }
 }
