+++ src/loader/index.ts
@@ -2,4 +2,4 @@ import "./SpineLoader";
 import "./SpineAtlasLoader";
 
 export { SpineResource } from "./SpineResource";
-export * from "./LoaderUtils";
+export { createSpineResource, createTextureAtlas } from "./LoaderUtils";
